Extract out new `Write.Era` module from `Write.Tx`
